# Apply optimizations in DeepSeek forward_absorb to forward_normal