<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation

In the past few weeks, there have been a lot of optimization PRs targeting at DeepSeek decoding with absorption, but they are still not applied to forward_normal/forward_normal_chunked_kv with MHA.

This PR applies these methods to forward_normal, so the performance of prefilling should increase a little bit.  What's more, the PR adds a new kernel fusion method that fuses the concat of q_nope/q_pe and k_nope/k_pe before mha.

**No more than 1% increase in prefill throughput, so not urgent for merging.**

<!-- Explain the purpose of this PR and the goals it aims to achieve. -->

## Modifications

- [x] Overlap qk norm/kv_b_proj (ref: #5977)
- [x] Apply set_mla_kv_buffer kernel (ref: #5748)
- [x] Fuse the concat of q_nope/q_pe and k_nope/k_pe before mha
- [x] Add accuracy/profiling/benchmark results

<!-- Describe the changes made in this PR. -->

## Accuracy Test

### Launch
```bash
python3 -m sglang.launch_server --model /dev/shm/DeepSeek-V3 --tp 8 --trust-remote-code --disable-radix
```

### GSM8K
```bash
python3 benchmark/gsm8k/bench_sglang.py --num-shots 8 --num-questions 1319 --parallel 1319
```

```bash
Accuracy: 0.944
Invalid: 0.000
Latency: 114.813 s
Output throughput: 1251.989 token/s
```

### MMLU

```bash
bash benchmark/mmlu/download_data.sh
python3 benchmark/mmlu/bench_sglang.py --nsub 100 --ntrain 5 --parallel 2000
```

```bash
Total latency: 377.314
Average accuracy: 0.876
```

## Benchmark
```bash
python3 -m sglang.bench_one_batch --model-path /dev/shm/DeepSeek-V3 --tp 8 --batch 1 --input-len 8192  --output-len 10 --disable-radix

# Main Branch
Prefill. latency: 0.38455 s, throughput:  21303.07 token/s
Decode 0. Batch size: 1, latency: 0.01322 s, throughput:     75.64 token/s
Decode 1. Batch size: 1, latency: 0.01263 s, throughput:     79.18 token/s
Decode 2. Batch size: 1, latency: 0.01263 s, throughput:     79.18 token/s
Decode 3. Batch size: 1, latency: 0.01259 s, throughput:     79.43 token/s
Decode 4. Batch size: 1, latency: 0.01261 s, throughput:     79.32 token/s
Decode.  median latency: 0.01263 s, median throughput:     79.21 token/s
Total. latency:  0.499 s, throughput:  16446.57 token/s

# PR
Prefill. latency: 0.38151 s, throughput:  21472.68 token/s
Decode 0. Batch size: 1, latency: 0.01345 s, throughput:     74.35 token/s
Decode 1. Batch size: 1, latency: 0.01267 s, throughput:     78.95 token/s
Decode 2. Batch size: 1, latency: 0.01267 s, throughput:     78.90 token/s
Decode 3. Batch size: 1, latency: 0.01270 s, throughput:     78.71 token/s
Decode 4. Batch size: 1, latency: 0.01266 s, throughput:     78.97 token/s
Decode.  median latency: 0.01267 s, median throughput:     78.95 token/s
Total. latency:  0.496 s, throughput:  16526.78 token/s
```

```bash
python3 -m sglang.bench_one_batch --model-path /dev/shm/DeepSeek-V3 --tp 8 --batch 16 --input-len 1024   --output-len 10 --disable-radix

# Main Branch
Prefill. latency: 0.60104 s, throughput:  27259.42 token/s
Decode 0. Batch size: 16, latency: 0.01834 s, throughput:    872.63 token/s
Decode 1. Batch size: 16, latency: 0.01765 s, throughput:    906.67 token/s
Decode 2. Batch size: 16, latency: 0.01769 s, throughput:    904.27 token/s
Decode 3. Batch size: 16, latency: 0.01761 s, throughput:    908.52 token/s
Decode 4. Batch size: 16, latency: 0.01750 s, throughput:    914.05 token/s
Decode.  median latency: 0.01765 s, median throughput:    906.31 token/s
Total. latency:  0.761 s, throughput:  21740.68 token/s

# PR
Prefill. latency: 0.59599 s, throughput:  27490.20 token/s
Decode 0. Batch size: 16, latency: 0.01805 s, throughput:    886.44 token/s
Decode 1. Batch size: 16, latency: 0.01764 s, throughput:    907.18 token/s
Decode 2. Batch size: 16, latency: 0.01812 s, throughput:    882.83 token/s
Decode 3. Batch size: 16, latency: 0.01813 s, throughput:    882.45 token/s
Decode 4. Batch size: 16, latency: 0.01806 s, throughput:    885.70 token/s
Decode.  median latency: 0.01780 s, median throughput:    899.05 token/s
Total. latency:  0.757 s, throughput:  21860.25 token/s

```


## Profile Result
```bash
python3 -m sglang.bench_one_batch --model-path /dev/shm/DeepSeek-V3 --tp 8 --cuda-graph-max-bs 32 --disable-radix --batch 16 --input-len 1024  --output-len 2 --profile
```

Main branch: 
![截屏2025-05-10 14 21 17](https://github.com/user-attachments/assets/6e987612-937d-4a62-87e6-9701d3ce8471)

This PR: 
![截屏2025-05-10 14 22 12](https://github.com/user-attachments/assets/f5a387e3-76a9-418b-93b0-6352027e6bc1)


## Checklist

- [ ] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [ ] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [ ] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.