@Fridge003 2025-05-12
> QQ why decoding is slower

I didn't change any code for decoding. The difference is really small, so it might be related to logic in `bench_one_batch.py`.
I will change for `bench_serving.py` and do more benchmarks later.